Major performance refactor with a few breaking changes
- Change how tables and primary keys are fetched to minimize reads
- Allow for more efficient bulk writing using Sqlite batches
- Rework abstractions to expose the Sqlite batch API
- Rename classes to better reflect their goals
- Correctly identify and forbid semicolon separated statements
