+++ lib/sqlite_crdt.dart
@@ -9,6 +9,7 @@ import 'package:sqflite_common_ffi_web/sqflite_ffi_web.dart';
 import 'package:sql_crdt/sql_crdt.dart';
 import 'package:sqlite_crdt/src/sqlite_api.dart';
 
+import 'src/batch_executor.dart';
 import 'src/is_web_locator.dart';
 
 export 'package:sqflite_common/sqlite_api.dart';
@@ -17,7 +18,7 @@ export 'package:sql_crdt/sql_crdt.dart';
 class SqliteCrdt extends SqlCrdt {
   final Database _db;
 
-  SqliteCrdt._(this._db) : super(SqliteApi(_db));
+  SqliteCrdt._(this._db) : super(ExecutorApi(_db));
 
   /// Open or create a SQLite container as a SqlCrdt instance.
   ///
@@ -27,8 +28,8 @@ class SqliteCrdt extends SqlCrdt {
     String path, {
     bool singleInstance = true,
     int? version,
-    FutureOr<void> Function(BaseCrdt crdt, int version)? onCreate,
-    FutureOr<void> Function(BaseCrdt crdt, int from, int to)? onUpgrade,
+    FutureOr<void> Function(Database db, int version)? onCreate,
+    FutureOr<void> Function(Database db, int from, int to)? onUpgrade,
   }) =>
       _open(path, false, singleInstance, version, onCreate, onUpgrade);
 
@@ -37,20 +38,31 @@ class SqliteCrdt extends SqlCrdt {
   static Future<SqliteCrdt> openInMemory({
     bool singleInstance = false,
     int? version,
-    FutureOr<void> Function(BaseCrdt crdt, int version)? onCreate,
-    FutureOr<void> Function(BaseCrdt crdt, int from, int to)? onUpgrade,
+    FutureOr<void> Function(CrdtTableExecutor db, int version)? onCreate,
+    FutureOr<void> Function(CrdtTableExecutor db, int from, int to)? onUpgrade,
   }) =>
-      _open(null, true, singleInstance, version, onCreate, onUpgrade);
-
-  Future<void> close() => _db.close();
+      _open(
+        null,
+        true,
+        singleInstance,
+        version,
+        onCreate == null
+            ? null
+            : (db, version) =>
+                onCreate(CrdtTableExecutor(ExecutorApi(db)), version),
+        onUpgrade == null
+            ? null
+            : (db, from, to) =>
+                onUpgrade(CrdtTableExecutor(ExecutorApi(db)), from, to),
+      );
 
   static Future<SqliteCrdt> _open(
     String? path,
     bool inMemory,
     bool singleInstance,
     int? version,
-    FutureOr<void> Function(BaseCrdt crdt, int version)? onCreate,
-    FutureOr<void> Function(BaseCrdt crdt, int from, int to)? onUpgrade,
+    FutureOr<void> Function(Database db, int version)? onCreate,
+    FutureOr<void> Function(Database db, int from, int to)? onUpgrade,
   ) async {
     if (sqliteCrdtIsWeb && !inMemory && path!.contains('/')) {
       path = path.substring(path.lastIndexOf('/') + 1);
@@ -68,13 +80,8 @@ class SqliteCrdt extends SqlCrdt {
       options: SqfliteOpenDatabaseOptions(
         singleInstance: singleInstance,
         version: version,
-        onCreate: onCreate == null
-            ? null
-            : (db, version) => onCreate.call(BaseCrdt(SqliteApi(db)), version),
-        onUpgrade: onUpgrade == null
-            ? null
-            : (db, from, to) =>
-                onUpgrade.call(BaseCrdt(SqliteApi(db)), from, to),
+        onCreate: onCreate,
+        onUpgrade: onUpgrade,
       ),
     );
 
@@ -82,4 +89,22 @@ class SqliteCrdt extends SqlCrdt {
     await crdt.init();
     return crdt;
   }
+
+  Future<void> close() => _db.close();
+
+  @override
+  Future<Iterable<String>> getTables() async => (await _db.rawQuery('''
+        SELECT name FROM sqlite_schema
+        WHERE type ='table' AND name NOT LIKE 'sqlite_%'
+      ''')).map((e) => e['name'] as String);
+
+  @override
+  Future<Iterable<String>> getTableKeys(String table) async =>
+      (await _db.rawQuery('''
+         SELECT name FROM pragma_table_info(?1)
+         WHERE pk > 0
+       ''', [table])).map((e) => e['name'] as String);
+
+  BatchExecutor batch() =>
+      BatchExecutor(_db.batch(), canonicalTime.increment());
 }
